split valueExecutor operations into separate files
