+++ server/src/graql/reasoner/cache/SemanticDifference.java
@@ -26,6 +26,7 @@ import grakn.core.concept.thing.Relation;
 import grakn.core.concept.type.Role;
 import grakn.core.concept.type.Type;
 import grakn.core.graql.executor.property.ValueExecutor;
+import grakn.core.graql.executor.property.value.Operation;
 import grakn.core.graql.reasoner.atom.predicate.ValuePredicate;
 import grakn.core.graql.reasoner.unifier.Unifier;
 import grakn.core.graql.reasoner.unifier.UnifierType;
@@ -113,7 +114,7 @@ public class SemanticDifference {
             return (type == null || type.subs().anyMatch(t -> t.equals(concept.asThing().type()))) &&
                     (role == null || role.subs().anyMatch(r -> r.equals(concept.asRole()))) &&
                     (vps.isEmpty() || vps.stream().allMatch(
-                            vp -> ValueExecutor.Operation.of(vp.getPredicate()).test(concept.asAttribute().value())
+                            vp -> Operation.of(vp.getPredicate()).test(concept.asAttribute().value())
                     ));
         });
     }
